requirements.txt
django>=3.0.0,<3.1.0
djangorestframework>=3.1,<4.0
social-auth-core>=3.0,<4.0
social-auth-app-django>=3.1,<4.0